+++ CHANGELOG.rst
@@ -2,9 +2,24 @@ Change Log
 =============
 
 
-[1.1.1] - (2023--) 
+[1.2.0] - (2023-11-20) 
 ----------------------
-- [ADDED]   geopackage as output format
+- [MOVED]   dave structure functions to DAVE client
+- [MOVED]   read simone function to DAVE client
+- [MOVED]   read gaslib function to DAVE client
+- [MOVED]   model module to DAVE client
+- [MOVED]   io module to DAVE client
+- [CHANGED] archiv i/o function in seperated file 
+- [MOVED]   plotting module to DAVE client
+- [ADDED]   building height based on raster data
+- [ADDED]   extend api with functions for database managment
+- [ADDED]   environment files for the possibility to install DAVE via mamba
+- [ADDED]   converter for the multiphysical network simulator MYNTS
+- [ADDED]   population data from census and the possibility to request the raster data
+- [ADDED]   importer for data from the gassimulation softwaretool SIMONE
+- [ADDED]   api restriction by user role
+- [ADDED]   option to choose year for nuts regions (2013, 2016, 2021)
+- [ADDED]   geopackage as possible output format
 - [ADDED]   extend geographical data with more landuse information and data for waterways
 - [CHANGED] input parameters for geographical objects reduced to one parameter "geodata"